requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.1
setuptools==65.5.0